Avoid exposing publishing_api globally via Service
We should isolate uses of the Publishing API to avoid making the
premature decision that each usage will have the same client settings.
For example, we may want draft actions to use a different timeout than
live actions, or for get_linkables to use a longer timeout.
